requirements.txt
Flask
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
